requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
email_validator==2.1.1
qrcode[pil]==7.4.2
Pillow==10.3.0
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.3
